Move StatsPackage to lib
